Finished app bar, additionally fixed getting cached recipe from database
