+++ makemelunch/package.json
@@ -10,11 +10,11 @@
     "export": "npm run build && next export -o _static"
   },
   "dependencies": {
-    "@chakra-ui/react": "^2.2.4",
     "@emotion/react": "^11.9.3",
     "@emotion/styled": "^11.9.3",
     "@mui/icons-material": "^5.8.4",
     "@mui/material": "^5.9.1",
+    "@mui/styles": "^5.9.1",
     "axios": "^0.27.2",
     "bootstrap": "^5.1.3",
     "firebase": "^9.9.0",
