+++ makemelunch/pages/ingredient/showRecipe.jsx
@@ -6,48 +6,18 @@ import { useState, useEffect } from "react";
 import axios from "axios";
 import { useAuth } from "../../context/UserAuthContext";
 import styles from "../../styles/ing.module.css";
-import { Container, Row, Button, Col } from "react-bootstrap";
 import { addRecipe, viewRecipe } from "../../helper/firebaseHelper";
+import Container from "@mui/material/Container";
+import MainAppBar from "../components/mainappbar";
+import { Button } from "@mui/material";
+import { Box } from "@mui/system";
 
 function ShowRecipe() {
   const auth = useAuth();
   const router = useRouter();
 
-  const handleClick = () => {
-    auth.logOut();
-  };
-
-  const headStyle = {
-    color: "green",
-    padding: "10px",
-    fontFamily: "Sans-Serif",
-    textAlign: "center",
-  };
-
   const editStyle = {
     padding: "10px",
-    textAlign: "center",
-  };
-
-  const buttonStyle = {
-    padding: "10px",
-    textAlign: "center",
-    color: "blue",
-  };
-
-  const optStyle = {
-    padding: "10px",
-    textAlign: "center",
-    margin: "20px",
-  };
-
-  const backStyle = {
-    color: "red",
-    margin: "30px",
-  };
-
-  const summaryStyle = {
-    color: "black",
     textAlign: "center",
   };
 
@@ -58,7 +28,6 @@ function ShowRecipe() {
     event.preventDefault();
 
     const data = router.query.id;
-    var summary;
     let options = {
       method: "GET",
       url: `https://spoonacular-recipe-food-nutrition-v1.p.rapidapi.com/recipes/${data}/information`,
@@ -68,7 +37,7 @@ function ShowRecipe() {
           "spoonacular-recipe-food-nutrition-v1.p.rapidapi.com",
       },
     };
-    if (viewRecipe(auth).id != data) {
+    if (data) {
       axios
         .request(options)
         .then(function (recipe) {
@@ -82,122 +51,70 @@ function ShowRecipe() {
           return;
         });
     } else {
-      setObject(viewRecipe(auth));
+      setObject(await viewRecipe(auth));
     }
   };
   if (!recipeObject) {
     return (
-      <main className={styles.main}>
+      <>
         <Head>
           <title>Show Recipe</title>
           <link rel="icon" href="/favicon.ico" />
         </Head>
-        <Row>
-          <h1 style={headStyle}>
-            <Link href="/">MakeMeLunch</Link>
-          </h1>
-        </Row>
-
-        <Row>
-          <Button style={optStyle}>
-            <Link href="/ingredient/viewingredients"> View Ingredients </Link>
-          </Button>
-          <Button style={optStyle}>
-            <Link href="/ingredient/addIng"> Add Ingredients </Link>
-          </Button>
-          <Button style={optStyle}>
-            <Link href="/ingredient/deleteIng"> Remove Ingredients </Link>
-          </Button>
-          <Button style={optStyle}>
-            <Link href="/ingredient/searchByIng">
-              Search for Recipe by Ingredient
-            </Link>
-          </Button>
-        </Row>
-        <Row>
-          <Button style={backStyle}>
-            <Link href="/">
-              <a onClick={() => handleClick()}>
-                {" "}
-                <b>Log Out </b>
-              </a>
-            </Link>
-          </Button>
-        </Row>
-        <br />
-        <form onSubmit={(event) => show(event)}>
-          <button type="submit">Show Recipe</button>
-        </form>
-      </main>
+        <MainAppBar />
+        <Container maxWidth="xl" sx={{ textAlign: "center" }}>
+          <Box component="form" onSubmit={show} noValidate>
+            <Button
+              type="submit"
+              variant="contained"
+              sx={{
+                mt: 3,
+                mb: 2,
+                backgroundColor: "green",
+                "&:hover": {
+                  backgroundColor: "#fff",
+                  color: "green",
+                },
+              }}
+            >
+              View Recipe
+            </Button>
+          </Box>
+        </Container>
+      </>
     );
   }
   if (recipeObject) {
     return (
       <div>
-        <Container fluid style={editStyle}>
-          <Head>
-            <title>Show Recipe</title>
-            <link rel="icon" href="/favicon.ico" />
-          </Head>
-          <Row>
-            <h1 style={headStyle}>
-              <Link href="/">MakeMeLunch</Link>
-            </h1>
-          </Row>
-
-          <Row>
-            <Button style={optStyle}>
-              <Link href="/ingredient/viewingredients"> View Ingredients </Link>
-            </Button>
-            <Button style={optStyle}>
-              <Link href="/ingredient/addIng"> Add Ingredients </Link>
-            </Button>
-            <Button style={optStyle}>
-              <Link href="/ingredient/deleteIng"> Remove Ingredients </Link>
-            </Button>
-            <Button style={optStyle}>
-              <Link href="/ingredient/searchByIng">
-                Search for Recipe by Ingredient
-              </Link>
-            </Button>
-          </Row>
-
-          <Row>
-            <Button style={backStyle}>
-              <Link href="/">
-                <a onClick={() => handleClick()}>
-                  {" "}
-                  <b>Log Out </b>
-                </a>
-              </Link>
-            </Button>
-          </Row>
-        </Container>
-        <Container fluid style={editStyle}>
-          <Col className="align-items-center">
-            <h1 className={styles.titlea}>{recipeObject.title}</h1>
-            <img src={recipeObject.image} alt={recipeObject.title} />
-            <h3 className={styles.optStyle}>Summary: </h3>
-            <div
-              className="product-des"
-              dangerouslySetInnerHTML={{ __html: recipeObject.summary }}
-            ></div>
-            <h3 className={styles.optStyle}>Ingredients: </h3>
-            <text>{recipeObject.extendedingredients}</text>
-            {recipeObject.extendedIngredients && (
-              <div>
-                {recipeObject.extendedIngredients?.map((ingredient, index) => (
-                  <div id={index} key={ingredient.id}>
-                    <h4>{ingredient.original}</h4>
-                  </div>
-                ))}
-              </div>
-            )}
-            <h3 className={styles.optStyle}>Instructions: </h3>
-            <text>{recipeObject.instructions}</text>
-            <h3 className={styles.optStyle}>Original Link: </h3>
-            <Link href={recipeObject.sourceUrl}> Link</Link>
-          </Col>
+        <Head>
+          <title>Show Recipe</title>
+          <link rel="icon" href="/favicon.ico" />
+        </Head>
+        <MainAppBar />
+        <Container style={editStyle}>
+          <h1 className={styles.titlea}>{recipeObject.title}</h1>
+          <img src={recipeObject.image} alt={recipeObject.title} />
+          <h3 className={styles.optStyle}>Summary: </h3>
+          <div
+            className="product-des"
+            dangerouslySetInnerHTML={{ __html: recipeObject.summary }}
+          ></div>
+          <h3 className={styles.optStyle}>Ingredients: </h3>
+          <text>{recipeObject.extendedingredients}</text>
+          {recipeObject.extendedIngredients && (
+            <div>
+              {recipeObject.extendedIngredients?.map((ingredient, index) => (
+                <div id={index} key={ingredient.id}>
+                  <h4>{ingredient.original}</h4>
+                </div>
+              ))}
+            </div>
+          )}
+          <h3 className={styles.optStyle}>Instructions: </h3>
+          <text>{recipeObject.instructions}</text>
+          <h3 className={styles.optStyle}>Original Link: </h3>
+          <Link href={recipeObject.sourceUrl}> Link</Link>
         </Container>
         <br />
         <br />
