+++ makemelunch/pages/ingredient/viewingredients.js
@@ -6,7 +6,8 @@ import { useState } from "react";
 //import styles from "../../styles/Home.module.css";
 import Link from "next/link";
 import styles from "../../styles/ing.module.css";
-import { Container, Row, Button } from "react-bootstrap";
+
+import MainAppBar from "../components/mainappbar.js";
 
 export default function ViewIngredients() {
   const auth = useAuth();
@@ -24,101 +25,34 @@ export default function ViewIngredients() {
     setIngs(str.split(","));
     setYes(1);
   };
-  const headStyle = {
-    color: "green",
-    padding: "10px",
-    fontFamily: "Sans-Serif",
-    textAlign: "center",
-  };
 
-  const editStyle = {
-    padding: "10px",
-    textAlign: "center",
-  };
-
-  const buttonStyle = {
-    padding: "10px",
-    textAlign: "center",
-    color: "blue",
-  };
-
-  const optStyle = {
-    padding: "10px",
-    textAlign: "center",
-    margin: "20px",
-  };
-
-  const backStyle = {
-    color: "red",
-    margin: "30px",
-  };
-
-  const summaryStyle = {
-    color: "black",
-    textAlign: "center",
-  };
   return (
     <>
-      <div className={styles.container}>
-        <Container fluid style={editStyle}>
-          <Head>
-            <title>View Ingredients</title>
-            <link rel="icon" href="/favicon.ico" />
-          </Head>
-          <Row>
-            <h1 style={headStyle}>
-              <Link href="/">MakeMeLunch</Link>
-            </h1>
-          </Row>
+      <Head>
+        <title>View Ingredients</title>
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <MainAppBar />
 
-          <Row>
-            <Button style={optStyle}>
-              <Link href="/ingredient/viewingredients"> View Ingredients </Link>
-            </Button>
-            <Button style={optStyle}>
-              <Link href="/ingredient/addIng"> Add Ingredients </Link>
-            </Button>
-            <Button style={optStyle}>
-              <Link href="/ingredient/deleteIng"> Remove Ingredients </Link>
-            </Button>
-            <Button style={optStyle}>
-              <Link href="/ingredient/searchByIng">
-                Search for Recipe by Ingredient
-              </Link>
-            </Button>
-          </Row>
-
-          <Row>
-            <Button style={backStyle}>
-              <Link href="/">
-                <a onClick={() => handleClick()}>
-                  {" "}
-                  <b>Log Out </b>
-                </a>
-              </Link>
-            </Button>
-          </Row>
-        </Container>
-        <main className={styles.main}>
-          <h1 className={styles.title}>View Ingredients</h1>
-          <br />
-          <form onSubmit={(event) => viewIng(event)}>
-            <button type="submit">View</button>
-          </form>
-          <br />
-          <Link href="javascript:history.back()">&larr; Go back</Link>
-          <br />
-          {yes && (
-            <div className={styles.viewingredients}>
-              {ings.map((ing, index) => (
-                <div id={index} key={ing}>
-                  <h3 id={ing}>{ing}</h3>
-                </div>
-              ))}
-            </div>
-          )}
-        </main>
-      </div>
+      <main className={styles.main}>
+        <h1 className={styles.title}>View Ingredients</h1>
+        <br />
+        <form onSubmit={(event) => viewIng(event)}>
+          <button type="submit">View</button>
+        </form>
+        <br />
+        <Link href="javascript:history.back()">&larr; Go back</Link>
+        <br />
+        {yes && (
+          <div className={styles.viewingredients}>
+            {ings.map((ing, index) => (
+              <div id={index} key={ing}>
+                <h3 id={ing}>{ing}</h3>
+              </div>
+            ))}
+          </div>
+        )}
+      </main>
     </>
   );
 }
